infill_subj_arithmetic/requirements.txt
torch==2.1.0
transformers==4.34.0
peft==0.5.0
wandb==0.15.12
rich==13.6.0
hydra-core==1.3.2
editdistance==0.6.2
bitsandbytes==0.41.1
tqdm==4.66.1
numpy==1.26.0
trl==0.7.1
scipy==1.11.3
ipykernel==6.25.2